dashboard/requirements.txt
flask==2.3.3
google-cloud-storage==2.3.0
gunicorn==21.2.0
jsonlines==4.0.0
python-dateutil>=2.8.2
Werkzeug>=2.3.0
Jinja2>=3.1.2
redis==5.0.7
rq==1.16.2
beautifulsoup4==4.12.3
google-cloud-bigquery==3.20.0
pyarrow>=3.0.0
firebase-admin==6.2.0
pydantic==2.5.0
flask-limiter==3.5.0
flask-swagger-ui==4.11.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-cloud-secret-manager==2.16.4
